% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_definitions.R
\name{get_definitions}
\alias{get_definitions}
\title{Get International Standard Classification of Occupations Definitions}
\usage{
get_definitions(lang = "eng")
}
\arguments{
\item{lang}{Un carácter indicando el lenguaje que queremos recuperar en formato
ISO 639-2 Alpha-3 (https://www.loc.gov/standards/iso639-2/php/code_list.php). This
parameter is 'eng' by default. Could be 'eng', 'spa', 'fre' or 'cat'.}
}
\value{
This functions returns a data.frame containing: the major group code (\code{major_code}),
the major group name (\code{major_name}), the sub-major code (\code{submajor_code}), the sub-major
name (\code{submajor_name}), minor group code (\code{minor_code}), minor group name (\code{minor_name}),
unit group code (\code{unit_code}), unit group name (\code{unit_name})
}
\description{
This functions allows the user to load a data.frame containing
all the ISCO definitions in multiple languages.
}
\examples{
get_definitions('spa')

}
