% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco_ui.R
\name{isco_ui}
\alias{isco_ui}
\title{International Standard Classification of Occupations UI}
\usage{
isco_ui(id)
}
\arguments{
\item{id}{The input slot that will be used to access the value}
}
\description{
This functions allows the user to build the ISCO user
interface in a shiny app.
}
\examples{
\dontrun{
## Only run examples in interactive R sessions
if(interactive()) {

options(device.ask.default = FALSE)

library(shiny)

ui <- fluidPage(

 sidebarLayout(

   sidebarPanel(
     width = 3,
     isco_ui("isco")
   ),

   mainPanel(
     tableOutput("tbl")
   )
 )

)

server <- function(input, output, session) {

 mod_data <- isco_server("isco", lang = 'spa')


 output$tbl <- renderTable({

   get_definitions(lang = 'spa') \%>\%
     filter(major_code == mod_data$major_code()) \%>\%
     filter(submajor_code == mod_data$submajor_code()) \%>\%
     filter(minor_code == mod_data$minor_code()) \%>\%
     filter(unit_code == mod_data$unit_code()) \%>\%
     select(
       major_name,
       submajor_name,
       minor_name,
       unit_name
     )

 })



}

shinyApp(ui, server)

}
}

}
